+++ Tests/Integration/DSCResources/Firefox.config.ps1
@@ -26,7 +26,7 @@ configuration Firefox_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]