+++ Tests/Integration/DSCResources/IisSite.config.ps1
@@ -26,7 +26,7 @@ configuration IisSite_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]