+++ Tests/Integration/DSCResources/McAfee.config.ps1
@@ -27,7 +27,7 @@ configuration McAfee_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]