+++ Tests/Integration/DSCResources/SqlServer.config.ps1
@@ -26,7 +26,7 @@ configuration SqlServer_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]