+++ Tests/Integration/DSCResources/Vsphere.config.ps1
@@ -35,7 +35,7 @@ configuration Vsphere_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [object]