+++ Tests/Integration/DSCResources/WindowsDefender.config.ps1
@@ -26,7 +26,7 @@ configuration WindowsDefender_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]