+++ Tests/Integration/DSCResources/WindowsDnsServer.config.ps1
@@ -26,7 +26,7 @@ configuration WindowsDnsServer_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]