+++ Tests/Integration/DSCResources/WindowsFirewall.config.ps1
@@ -26,7 +26,7 @@ configuration WindowsFirewall_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]