+++ Tests/Integration/DSCResources/WindowsServer.config.ps1
@@ -26,7 +26,7 @@ configuration WindowsServer_config
 
         [Parameter()]
         [string[]]
-        $SkipRuleCategory,
+        $SkipRuleSeverity,
 
         [Parameter()]
         [hashtable]