+++ source/DSCResources/FireFox/FireFox.schema.psm1
@@ -61,15 +61,14 @@ configuration FireFox
         $SkipRuleType,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [ValidateSet('CAT_I', 'CAT_II', 'CAT_III')]
         [string[]]
-        $SkipRuleCategory
+        $SkipRuleSeverity
     )
 
     ##### BEGIN DO NOT MODIFY #####
     $stig = [STIG]::New('FireFox', 'All', $StigVersion)
-    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleCategory)
+    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleSeverity)
     ##### END DO NOT MODIFY #####
 
     Import-DscResource -ModuleName FileContentDsc -ModuleVersion 1.3.0.151