+++ source/DSCResources/IisSite/IisSite.schema.psm1
@@ -77,15 +77,14 @@ configuration IisSite
         $SkipRuleType,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [ValidateSet('CAT_I', 'CAT_II', 'CAT_III')]
         [string[]]
-        $SkipRuleCategory
+        $SkipRuleSeverity
     )
 
     ##### BEGIN DO NOT MODIFY #####
     $stig = [STIG]::New('IISSite', $IisVersion, $StigVersion)
-    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleCategory)
+    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleSeverity)
     ##### END DO NOT MODIFY #####
 
     Import-DscResource -ModuleName PSDSCresources -ModuleVersion 2.12.0.0