+++ source/DSCResources/InternetExplorer/InternetExplorer.schema.psm1
@@ -64,15 +64,14 @@ configuration InternetExplorer
         $SkipRuleType,
 
         [Parameter()]
-        [ValidateNotNullOrEmpty()]
         [ValidateSet('CAT_I', 'CAT_II', 'CAT_III')]
         [string[]]
-        $SkipRuleCategory
+        $SkipRuleSeverity
     )
 
     ##### BEGIN DO NOT MODIFY #####
     $stig = [STIG]::New('InternetExplorer', $BrowserVersion, $StigVersion)
-    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleCategory)
+    $stig.LoadRules($OrgSettings, $Exception, $SkipRule, $SkipRuleType, $SkipRuleSeverity)
     ##### END DO NOT MODIFY #####
 
     Import-DscResource -ModuleName GPRegistryPolicyDsc -ModuleVersion 1.2.0